Major refactor. Created type alias for better documentation, added an interface for KParser for documentation and creating more flexible testing. Scattered cleanup and reorganization.
finished documentation and added a base set of unit tests for KParsing.

wip - attempted pull out of TornadoFX configurations

Looks like I've successfully pulled out TornadoFX configurations? Will need to allow custom configurations at the creation of KParser and make TornadoFX specific configurations not independent functions... maybe an object

wip: application is running but there is a bug where only one node is counted in the digraph

further refactoring.  MappedClassNodes appears to be erased for every iteration of the class set, so I will need to figure out how to preserve that hashmap

fixed the bug - I had been initializing KParser at every file *facepalm*
