--- src/main/kotlin/com/github/ast/parser/FrameworkConfigurations.kt
@@ -1,9 +1,0 @@
-package com.github.ast.parser
-
-sealed class Views(var viewClass: String? = "", var viewType: String? = "")
-
-data class TornadoFXView(
-        var view: String? = "",
-        var type: String? = "",
-        var scope: String? = ""
-) : Views(view, type)
